+++ src/mjolnir/pbfgraphparser.cc
@@ -175,6 +175,7 @@ struct graph_parser {
     include_platforms_ = pt.get<bool>("include_platforms", false);
     include_driveways_ = pt.get<bool>("include_driveways", true);
     include_construction_ = pt.get<bool>("include_construction", false);
+    pedestrian_areas_ = pt.get<bool>("pedestrian_areas", false);
     infer_internal_intersections_ =
         pt.get<bool>("data_processing.infer_internal_intersections", true);
     infer_turn_channels_ = pt.get<bool>("data_processing.infer_turn_channels", true);
@@ -374,6 +375,11 @@ struct graph_parser {
     tag_handlers_["amenity"] = [this]() {
       if (tag_.second == "yes") {
         amenity_ = tag_.second;
+      }
+    };
+    tag_handlers_["pedestrian_area"] = [this]() {
+      if (pedestrian_areas_) {
+        way_.set_area(tag_.second == "true" ? true : false);
       }
     };
 
@@ -3840,6 +3846,7 @@ struct graph_parser {
     uint64_t from_way_id = 0;
     bool isRestriction = false, isTypeRestriction = false, hasRestriction = false;
     bool isRoad = false, isRoute = false, isBicycle = false, isConnectivity = false;
+    bool isMultipolygon = false, isPedestrian = false, isArea = false;
     bool isConditional = false, isProbable = false, has_multiple_times = false;
     uint32_t bike_network_mask = 0;
 
@@ -3857,6 +3864,8 @@ struct graph_parser {
           isRoute = true;
         } else if (tag.second == "connectivity") {
           isConnectivity = true;
+        } else if (tag.second == "multipolygon") {
+          isMultipolygon = true;
         }
       } else if (tag.first == "route") {
         if (tag.second == "road") {
@@ -3983,6 +3992,14 @@ struct graph_parser {
         to = tag.second;
       } else if (tag.first == "from") {
         from = tag.second;
+      } else if (tag.first == "highway") {
+        if (tag.second == "pedestrian") {
+          isPedestrian = true;
+        }
+      } else if (tag.first == "area") {
+        if (tag.second == "yes") {
+          isArea = true;
+        }
       }
     } // for (const auto& tag : results)
 
@@ -4264,6 +4281,19 @@ struct graph_parser {
           complex_restrictions_from_->push_back(restriction);
         } else { // simple restriction
           osmdata_.restrictions.insert(RestrictionsMultiMap::value_type(from_way_id, restriction));
+        }
+      }
+    } else if (isMultipolygon && isPedestrian && isArea && pedestrian_areas_) {
+      for (const auto& member : members) {
+        OSMAreaMember area_member;
+        if (member.role == "outer" && member.member_type == osmium::item_type::way) {
+          area_member.is_outer = true;
+          area_member.way_id = member.member_id;
+          osmdata_.area_relations.insert(AreaMultiMap::value_type(osmid, area_member));
+        } else if (member.role == "inner" && member.member_type == osmium::item_type::way) {
+          area_member.is_outer = false;
+          area_member.way_id = member.member_id;
+          osmdata_.area_relations.insert(AreaMultiMap::value_type(osmid, area_member));
         }
       }
     }
@@ -5033,6 +5063,9 @@ struct graph_parser {
   // Configuration option indicating whether or not to infer turn channels during the graph
   // enhancer phase or use the turn_channel key from the pbf
   bool infer_turn_channels_;
+
+  // Configuration option indicating whether or not to generate edges and route trough pedestrian areas
+  bool pedestrian_areas_;
 
   // Configuration option indicating whether or not to process the direction key on the ways or
   // utilize the guidance relation tags during the parsing phase